requirements.txt
chromadb==0.5.5
openai==1.46.0
python-dotenv==1.0.1